requirements.txt
asttokens==3.0.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
fonttools==4.61.1
ImageIO==2.37.2
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapy==1.2.5
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
scikit-image==0.26.0
scipy==1.16.3
six==1.17.0
stack-data==0.6.3
tifffile==2025.12.20
traitlets==5.14.3
wcwidth==0.2.14
